Development

+++ app/Datamapper/image.js
@@ -3,31 +3,25 @@ const { ApiError } = require('../errors/apiErrors');
 
 
 module.exports = {
-  async addImage(workspaceId, imageList) {
+  async addImage(workspaceId, mainImage, otherImagesLink) {
     
     const queryImage = [];
     const values = [];
     let counter = 2;
 
-    for(const key of imageList){
-      if (key.fieldname.includes("mainImage")) {
-        queryImage.push(`($1, $${counter}, true)`);
-      } else {
+    if (otherImagesLink[0].length > 0) {
+      for (const link of otherImagesLink) {
         queryImage.push(`($1, $${counter}, false)`);
+        values.push(link);
+        counter++;
       }
-
-      // Supprime la première partie du lien => public
-      const path = key.path
-      .split('/')
-      .slice(1)
-      .join('/');
-
-      values.push(path);
-      counter++;
     }
 
-    const queryString = `
-    INSERT INTO image (workspace_id, link, main_image) 
+    queryImage.push(`($1, $${counter}, true)`)
+    values.push(mainImage);
+
+    let queryString = `
+     INSERT INTO image (workspace_id, link, main_image) 
     VALUES ${queryImage.join(',')};
     `;
 
@@ -49,35 +43,26 @@ module.exports = {
     return result.rows
   },
 
-    async updateMainImage(workspaceId, imageList) {
+    async updateImage(workspaceId, body) {
+    let path = '';
+    let isMainImage = false;
 
-    await client.query(`DELETE FROM image WHERE workspace_id = $1 AND main_image = true`, [workspaceId]);
+    
+    if (body?.workspace_mainImage){
+      await client.query(`DELETE FROM image WHERE workspace_id = $1 AND main_image = true`, [workspaceId]);
+      path = body.workspace_mainImage;
+      isMainImage = true
 
-    // Supprime la première partie du lien => public
-    const path = imageList[0].path
-    .split('/')
-    .slice(1)
-    .join('/');
+    } else if (body?.workspace_image) {
+      path = body.workspace_image;
+      isMainImage = false;
+    }
 
-    await client.query(`INSERT INTO image (workspace_id, main_image, link) VALUES ($1, $2, $3)`, [workspaceId, true, path])
+    await client.query(`INSERT INTO image (workspace_id, main_image, link) VALUES ($1, $2, $3)`, [workspaceId, isMainImage, path])
 
     const result = await client.query(`SELECT id as image_id, link, main_image, workspace_id FROM image WHERE workspace_id = $1`, [workspaceId]);
 
     return result.rows;
 
   },
-
-  async updateImages(workspaceId, imageList) {
-
-    const path = imageList[0].path
-    .split('/')
-    .slice(1)
-    .join('/');
-
-    await client.query(`INSERT INTO image (workspace_id, main_image, link) VALUES ($1, $2, $3)`, [workspaceId, false, path]);
-
-    const result = await client.query(`SELECT id as image_id, link, main_image, workspace_id FROM image WHERE workspace_id = $1`, [workspaceId]);
-
-    return result.rows;
-  }
 };
